pyproject.toml
[project]
name = "nuvie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.13",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "structlog>=25.4.0",
]
